+++ api/application.ts
@@ -1,3 +1,4 @@
+import * as healthcheck from '@hmcts/nodejs-healthcheck'
 import * as bodyParser from 'body-parser'
 import * as cookieParser from 'cookie-parser'
 import * as express from 'express'
@@ -5,11 +6,17 @@ import * as helmet from 'helmet'
 import {getXuiNodeMiddleware} from './auth'
 import { getConfigValue, showFeature } from './configuration'
 import {
-    FEATURE_HELMET_ENABLED,
-    HELMET, PROTOCOL,
-    SESSION_SECRET,
+  FEATURE_HELMET_ENABLED, FEATURE_REDIS_ENABLED, FEATURE_TERMS_AND_CONDITIONS_ENABLED,
+  HELMET,
+  PROTOCOL, SERVICE_S2S_PATH,
+  SERVICES_CCD_COMPONENT_API_PATH,
+  SERVICES_CCD_DATA_STORE_API_PATH,
+  SERVICES_DOCUMENTS_API_PATH,
+  SERVICES_EM_ANNO_API_URL,
+  SERVICES_IDAM_API_URL, SERVICES_IDAM_LOGIN_URL,
+  SERVICES_TERMS_AND_CONDITIONS_URL,
+  SESSION_SECRET,
 } from './configuration/references'
-import * as health from './health'
 import * as log4jui from './lib/log4jui'
 import {JUILogger} from './lib/models'
 import * as tunnel from './lib/tunnel'
@@ -23,7 +30,7 @@ export const app = express()
 /**
  * Add Reform Standard health checks.
  */
-health.addReformHealthCheck(app)
+// health.addReformHealthCheck(app)
 
 if (showFeature(FEATURE_HELMET_ENABLED)) {
     app.use(helmet(getConfigValue(HELMET)))
@@ -33,6 +40,43 @@ app.use(cookieParser(getConfigValue(SESSION_SECRET)))
 
 // TODO: remove tunnel and configurations
 tunnel.init()
+
+export const checkServiceHealth = service => healthcheck.web(`${service}/health`, {
+  deadline: 6000,
+  timeout: 6000,
+})
+
+const healthChecks = {
+  checks: {
+    ccdComponentApi: checkServiceHealth(getConfigValue(SERVICES_CCD_COMPONENT_API_PATH)),
+    ccdDataApi: checkServiceHealth(getConfigValue(SERVICES_CCD_DATA_STORE_API_PATH)),
+    documentsApi: checkServiceHealth(getConfigValue(SERVICES_DOCUMENTS_API_PATH)),
+    emmoApi: checkServiceHealth(getConfigValue(SERVICES_EM_ANNO_API_URL)),
+    idamApi: checkServiceHealth(getConfigValue(SERVICES_IDAM_LOGIN_URL)),
+    idamWeb: checkServiceHealth(getConfigValue(SERVICES_IDAM_API_URL)),
+    s2s: checkServiceHealth(getConfigValue(SERVICE_S2S_PATH)),
+    // termsAndConditions: checkServiceHealth(getConfigValue(SERVICES_TERMS_AND_CONDITIONS_URL)),
+  },
+}
+
+if (showFeature(FEATURE_TERMS_AND_CONDITIONS_ENABLED)) {
+  healthChecks.checks = {...healthChecks.checks, ...{
+    termsAndConditions: checkServiceHealth(getConfigValue(SERVICES_TERMS_AND_CONDITIONS_URL)),
+  }}
+}
+
+if (showFeature(FEATURE_REDIS_ENABLED)) {
+  healthChecks.checks = {...healthChecks.checks, ...{
+    redis: healthcheck.raw(() => {
+      return app.locals.redisClient.connected ? healthcheck.up() : healthcheck.down()
+    }),
+  }}
+}
+
+console.log('healthChecks', healthChecks)
+
+healthcheck.addTo(app, healthChecks)
+
 app.use(getXuiNodeMiddleware())
 
 // applyProxy needs to be used before bodyParser
